Refactored new websockets impl into own modules Incorporated other suggestions made by team
