+++ sanic/server/websockets/impl.py
@@ -1,247 +1,17 @@
+import asyncio
 import random
 import struct
-from email.utils import formatdate
-from typing import (
-    Any,
-    Awaitable,
-    Callable,
-    Dict,
-    List,
-    MutableMapping,
-    Optional,
-    Union,
-    Iterable,
-    Sequence,
-    Mapping, AsyncIterator,
-    TYPE_CHECKING
-)
-import codecs
+from typing import Optional, Mapping, Iterable, Union, AsyncIterator, Sequence, Dict
 
-from httptools import HttpParserUpgrade  # type: ignore
+from websockets.connection import CLOSED, OPEN, Event, CLOSING
+from websockets.exceptions import ConnectionClosedError, ConnectionClosed
+from websockets.frames import prepare_ctrl, Frame, OP_PONG
 from websockets.server import ServerConnection
-from websockets.connection import Event, OPEN, CLOSING, CLOSED
-from websockets.exceptions import ConnectionClosed, InvalidHandshake, InvalidOrigin, InvalidUpgrade, \
-    ConnectionClosedError
 from websockets.typing import Data
-from websockets.frames import Frame, Opcode, prepare_ctrl, OP_PONG
-from websockets.utils import accept_key
 
-from sanic.exceptions import InvalidUsage, Forbidden, SanicException
-from sanic.server import HttpProtocol
+from sanic.log import error_logger
 from sanic.server.protocols.base_protocol import SanicProtocol
-from sanic.response import BaseHTTPResponse
-from sanic.log import error_logger, logger
-
-import asyncio
-
-ASIMessage = MutableMapping[str, Any]
-UTF8Decoder = codecs.getincrementaldecoder("utf-8")
-
-class WebsocketFrameAssembler:
-    """
-    Assemble a message from frames.
-    Code borrowed from aaugustin/websockets project:
-    https://github.com/aaugustin/websockets/blob/6eb98dd8fa5b2c896b9f6be7e8d117708da82a39/src/websockets/sync/messages.py
-    """
-    __slots__ = ("protocol", "read_mutex", "write_mutex", "message_complete", "message_fetched", "get_in_progress", "decoder", "completed_queue", "chunks", "chunks_queue", "paused", "get_id", "put_id")
-    if TYPE_CHECKING:
-        protocol: "WebsocketImplProtocol"
-        read_mutex: asyncio.Lock
-        write_mutex: asyncio.Lock
-        message_complete: asyncio.Event
-        message_fetched: asyncio.Event
-        completed_queue: asyncio.Queue
-        get_in_progress: bool
-        decoder: Optional[codecs.IncrementalDecoder]
-        # For streaming chunks rather than messages:
-        chunks: List[Data]
-        chunks_queue: Optional[asyncio.Queue[Optional[Data]]]
-        paused: bool
-
-
-    def __init__(self, protocol) -> None:
-
-        self.protocol = protocol
-
-        self.read_mutex = asyncio.Lock()
-        self.write_mutex = asyncio.Lock()
-
-        self.completed_queue = asyncio.Queue(maxsize=1)  # type: asyncio.Queue[Data]
-
-
-        # put() sets this event to tell get() that a message can be fetched.
-        self.message_complete = asyncio.Event()
-        # get() sets this event to let put()
-        self.message_fetched = asyncio.Event()
-
-        # This flag prevents concurrent calls to get() by user code.
-        self.get_in_progress = False
-
-        # Decoder for text frames, None for binary frames.
-        self.decoder = None
-
-        # Buffer data from frames belonging to the same message.
-        self.chunks = []
-
-        # When switching from "buffering" to "streaming", we use a thread-safe
-        # queue for transferring frames from the writing thread (library code)
-        # to the reading thread (user code). We're buffering when chunks_queue
-        # is None and streaming when it's a Queue. None is a sentinel
-        # value marking the end of the stream, superseding message_complete.
-
-        # Stream data from frames belonging to the same message.
-        self.chunks_queue = None
-
-        # Flag to indicate we've paused the protocol
-        self.paused = False
-
-
-    async def get(self, timeout: Optional[float] = None) -> Optional[Data]:
-        """
-        Read the next message.
-        :meth:`get` returns a single :class:`str` or :class:`bytes`.
-        If the :message was fragmented, :meth:`get` waits until the last frame
-        is received, then it reassembles the message.
-        If ``timeout`` is set and elapses before a complete message is
-        received, :meth:`get` returns ``None``.
-        """
-        async with self.read_mutex:
-            if timeout is not None and timeout <= 0:
-                if not self.message_complete.is_set():
-                    return None
-            assert not self.get_in_progress
-            self.get_in_progress = True
-
-            # If the message_complete event isn't set yet, release the lock to
-            # allow put() to run and eventually set it.
-            # Locking with get_in_progress ensures only one thread can get here.
-            if timeout is None:
-                completed = await self.message_complete.wait()
-            elif timeout <= 0:
-                completed = self.message_complete.is_set()
-            else:
-                completed = await asyncio.wait_for(self.message_complete.wait(), timeout=timeout)
-
-            # Unpause the transport, if its paused
-            if self.paused:
-                self.protocol.resume_frames()
-                self.paused = False
-            assert self.get_in_progress
-            self.get_in_progress = False
-
-            # Waiting for a complete message timed out.
-            if not completed:
-                return None
-
-            assert self.message_complete.is_set()
-            self.message_complete.clear()
-
-            joiner: Data = b"" if self.decoder is None else ""
-            # mypy cannot figure out that chunks have the proper type.
-            message: Data = joiner.join(self.chunks)  # type: ignore
-
-            assert not self.message_fetched.is_set()
-            self.message_fetched.set()
-            self.chunks = []
-            assert self.chunks_queue is None
-
-            return message
-
-    async def get_iter(self) -> AsyncIterator[Data]:
-        """
-        Stream the next message.
-        Iterating the return value of :meth:`get_iter` yields a :class:`str`
-        or :class:`bytes` for each frame in the message.
-        """
-        async with self.read_mutex:
-            assert not self.get_in_progress
-            self.get_in_progress = True
-
-            chunks = self.chunks
-            self.chunks = []
-            self.chunks_queue = asyncio.Queue()
-
-            # Sending None in chunk_queue supersedes setting message_complete
-            # when switching to "streaming". If message is already complete
-            # when the switch happens, put() didn't send None, so we have to.
-            if self.message_complete.is_set():
-                await self.chunks_queue.put(None)
-
-            # Locking with get_in_progress ensures only one thread can get here.
-            for c in chunks:
-                yield c
-            while True:
-                chunk = await self.chunks_queue.get()
-                if chunk is None:
-                    break
-                yield chunk
-
-            # Unpause the transport, if its paused
-            if self.paused:
-                self.protocol.resume_frames()
-                self.paused = False
-            assert self.get_in_progress
-            self.get_in_progress = False
-            assert self.message_complete.is_set()
-            self.message_complete.clear()
-
-            assert not self.message_fetched.is_set()
-
-            self.message_fetched.set()
-
-            assert self.chunks == []
-            self.chunks_queue = None
-
-    async def put(self, frame: Frame) -> None:
-        """
-        Add ``frame`` to the next message.
-        When ``frame`` is the final frame in a message, :meth:`put` waits
-        until the message is fetched, either by calling :meth:`get` or by
-        iterating the return value of :meth:`get_iter`.
-        :meth:`put` assumes that the stream of frames respects the protocol.
-        If it doesn't, the behavior is undefined.
-        """
-        #id = self.put_id
-        #self.put_id += 1
-        async with self.write_mutex:
-            if frame.opcode is Opcode.TEXT:
-                self.decoder = UTF8Decoder(errors="strict")
-            elif frame.opcode is Opcode.BINARY:
-                self.decoder = None
-            elif frame.opcode is Opcode.CONT:
-                pass
-            else:
-                # Ignore control frames.
-                return
-            data: Data
-            if self.decoder is not None:
-                data = self.decoder.decode(frame.data, frame.fin)
-            else:
-                data = frame.data
-            if self.chunks_queue is None:
-                self.chunks.append(data)
-            else:
-                await self.chunks_queue.put(data)
-
-            if not frame.fin:
-                return
-            if not self.get_in_progress:
-                self.paused = self.protocol.pause_frames()
-            # Message is complete. Wait until it's fetched to return.
-
-            if self.chunks_queue is not None:
-                await self.chunks_queue.put(None)
-
-            assert not self.message_complete.is_set()
-            self.message_complete.set()
-            assert not self.message_fetched.is_set()
-
-            # Release the lock to allow get() to run and eventually set the event.
-            await self.message_fetched.wait()
-            assert self.message_fetched.is_set()
-            self.message_fetched.clear()
-            self.decoder = None
-
+from .frame import WebsocketFrameAssembler
 
 class WebsocketImplProtocol:
     connection: ServerConnection
@@ -251,6 +21,7 @@ class WebsocketImplProtocol:
     ping_interval: Optional[float]
     ping_timeout: Optional[float]
     assembler: WebsocketFrameAssembler
+    pings: Dict[bytes, asyncio.Future]  # Dict[bytes, asyncio.Future[None]]
     pings: Dict[bytes, asyncio.Future]  # Dict[bytes, asyncio.Future[None]]
     conn_mutex: asyncio.Lock
     recv_lock: asyncio.Lock
@@ -334,7 +105,9 @@ class WebsocketImplProtocol:
         """
         if self.connection_lost_waiter is None:
             return False
-        if not self.connection_lost_waiter.done():
+        if self.connection_lost_waiter.done():
+            return True
+        else:
             try:
                 await asyncio.wait_for(
                     asyncio.shield(self.connection_lost_waiter), timeout
@@ -370,7 +143,7 @@ class WebsocketImplProtocol:
                     ping.set_result(None)
                 if ping_id == frame.data:
                     break
-            else:  # pragma: no cover
+            else:  # noqa
                 assert False, "ping_id is in self.pings"
             # Remove acknowledged pings from self.pings.
             for ping_id in ping_ids:
@@ -410,7 +183,7 @@ class WebsocketImplProtocol:
             raise
         except ConnectionClosed:
             pass
-        except BaseException:
+        except Exception:
             error_logger.warning("Unexpected exception in keepalive ping task")
 
     def fail_connection(self, code: int = 1006, reason: str = "") -> bool:
@@ -443,12 +216,17 @@ class WebsocketImplProtocol:
                 self.connection.fail_connection(code, reason)
                 for frame_data in self.connection.data_to_send():
                     self.io_proto.transport.write(frame_data)
+        if code == 1006:
+            # Special case: 1006 consider the transport already closed
+            self.connection.set_state(CLOSED)
         if self.close_connection_task is not None and not self.close_connection_task.done():
             if self.data_finished_fut is not None and not self.data_finished_fut.done():
                 self.data_finished_fut.cancel()
             # Don't close, auto_close_connection will take care of it.
             return False
-        SanicProtocol.close(self.io_proto)
+
+        # No auto-closer available, just abort the connection
+        SanicProtocol.close(self.io_proto, timeout=1.0)
         return True
 
     async def auto_close_connection(self) -> None:
@@ -485,14 +263,12 @@ class WebsocketImplProtocol:
             if self.connection_lost_waiter is not None and self.connection_lost_waiter.done():
                 if self.io_proto.transport is None or self.io_proto.transport.is_closing():
                     return
-            SanicProtocol.close(self.io_proto)
+            SanicProtocol.close(self.io_proto, timeout=999)
             if self.connection_lost_waiter is not None:
-                await self.wait_for_connection_lost()
-                if self.connection_lost_waiter.done():
+                if await self.wait_for_connection_lost(timeout=5):
                     return
                 error_logger.warning("Timeout waiting for TCP connection to close. Aborting")
-                if self.io_proto.transport is not None:
-                    self.io_proto.transport.abort()
+                SanicProtocol.abort(self.io_proto)
 
     def abort_pings(self) -> None:
         """
@@ -545,14 +321,14 @@ class WebsocketImplProtocol:
             :meth:`recv_streaming` concurrently
         """
 
-        # TODO HANDLE THE SITUATION WHERE THE CONNECTION IS CLOSED
-
         if self.recv_lock.locked():
             raise RuntimeError(
                 "cannot call recv while another task "
                 "is already waiting for the next message"
             )
         await self.recv_lock.acquire()
+        if self.connection.state in (CLOSED, CLOSING):
+            raise RuntimeError("Cannot receive from websocket interface after it is closed.")
         try:
             return await self.assembler.get(timeout)
         finally:
@@ -575,15 +351,14 @@ class WebsocketImplProtocol:
             :meth:`recv_streaming` concurrently
         """
 
-        # TODO HANDLE THE SITUATION WHERE THE CONNECTION IS CLOSED
-
         if self.recv_lock.locked():
             raise RuntimeError(
                 "cannot call recv_burst while another task "
                 "is already waiting for the next message"
             )
         await self.recv_lock.acquire()
-
+        if self.connection.state in (CLOSED, CLOSING):
+            raise RuntimeError("Cannot receive from websocket interface after it is closed.")
         messages = []
         try:
             # Prevent pausing the transport when we're
@@ -621,6 +396,8 @@ class WebsocketImplProtocol:
                 "is already waiting for the next message"
             )
         await self.recv_lock.acquire()
+        if self.connection.state in (CLOSED, CLOSING):
+            raise RuntimeError("Cannot receive from websocket interface after it is closed.")
         try:
             self.can_pause = False
             async for m in self.assembler.get_iter():
@@ -649,7 +426,10 @@ class WebsocketImplProtocol:
         """
         async with self.conn_mutex:
 
-            # TODO HANDLE THE SITUATION WHERE THE CONNECTION IS CLOSED
+            if self.connection.state in (CLOSED, CLOSING):
+                raise RuntimeError("Cannot write to websocket interface after it is closed.")
+            if self.data_finished_fut is None or self.data_finished_fut.cancelled() or self.data_finished_fut.done():
+                raise RuntimeError("Cannot write to websocket interface after it is finished.")
 
             # Unfragmented message -- this case must be handled first because
             # strings and bytes-like objects are iterable.
@@ -687,9 +467,8 @@ class WebsocketImplProtocol:
         (which will be encoded to UTF-8) or a bytes-like object.
         """
         async with self.conn_mutex:
-
-            # TODO HANDLE THE SITUATION WHERE THE CONNECTION IS CLOSED
-
+            if self.connection.state in (CLOSED, CLOSING):
+                raise RuntimeError("Cannot send a ping when the websocket interface is closed.")
             if data is not None:
                 data = prepare_ctrl(data)
 
@@ -716,7 +495,9 @@ class WebsocketImplProtocol:
         be a string (which will be encoded to UTF-8) or a bytes-like object.
         """
         async with self.conn_mutex:
-            # TODO HANDLE THE SITUATION WHERE THE CONNECTION IS CLOSED
+            if self.connection.state in (CLOSED, CLOSING):
+                # Cannot send pong after transport is shutting down
+                return
 
             data = prepare_ctrl(data)
 
@@ -729,17 +510,17 @@ class WebsocketImplProtocol:
                 await self.io_proto.send(data)
             else:
                 # Send an EOF
-                # We don't actually send it, just close the connection
+                # We don't actually send it, just trigger to autoclose the connection
                 if self.close_connection_task is not None and not self.close_connection_task.done() and \
                         self.data_finished_fut is not None and not self.data_finished_fut.done():
                     # Auto-close the connection
                     self.data_finished_fut.set_result(None)
                 else:
                     # This will fail the connection appropriately
-                    self.io_proto.close()
+                    SanicProtocol.close(self.io_proto, timeout=1.0)
 
     async def async_data_received(self, data_to_send, events_to_process):
-        if len(data_to_send) > 0:
+        if self.connection.state == OPEN and len(data_to_send) > 0:
             # receiving data can generate data to send (eg, pong for a ping)
             # send connection.data_to_send()
             await self.send_data(data_to_send)
@@ -756,7 +537,8 @@ class WebsocketImplProtocol:
     async def async_eof_received(self, data_to_send, events_to_process):
         # receiving EOF can generate data to send
         # send connection.data_to_send()
-        await self.send_data(data_to_send)
+        if self.connection.state == OPEN:
+            await self.send_data(data_to_send)
         if len(events_to_process) > 0:
             await self.process_events(events_to_process)
 
@@ -766,7 +548,7 @@ class WebsocketImplProtocol:
             self.data_finished_fut.set_result(None)
         else:
             # This will fail the connection appropriately
-            self.io_proto.close()
+            SanicProtocol.close(self.io_proto, timeout=1.0)
 
     def eof_received(self) -> Optional[bool]:
         self.connection.receive_eof()
@@ -784,167 +566,3 @@ class WebsocketImplProtocol:
         self.abort_pings()
         if self.connection_lost_waiter is not None:
             self.connection_lost_waiter.set_result(None)
-
-
-class WebSocketProtocol(HttpProtocol):
-    def __init__(
-        self,
-        *args,
-        websocket_timeout=10,
-        websocket_max_size=None,
-        websocket_max_queue=None,
-        websocket_read_limit=2 ** 16,
-        websocket_write_limit=2 ** 16,
-        websocket_ping_interval=20,
-        websocket_ping_timeout=20,
-        **kwargs,
-    ):
-        super().__init__(*args, **kwargs)
-        self.websocket = None # type: Union[None, WebsocketImplProtocol]
-        # self.app = None
-        self.websocket_timeout = websocket_timeout
-        self.websocket_max_size = websocket_max_size
-        if websocket_max_queue is not None and int(websocket_max_queue) > 0:
-            error_logger.warning(DeprecationWarning("websocket_max_queue is no longer used. No websocket message queueing is implemented."))
-        self.websocket_read_limit = websocket_read_limit
-        self.websocket_write_limit = websocket_write_limit
-        self.websocket_ping_interval = websocket_ping_interval
-        self.websocket_ping_timeout = websocket_ping_timeout
-
-    def connection_lost(self, exc):
-        if self.websocket is not None:
-            self.websocket.connection_lost(exc)
-        super().connection_lost(exc)
-
-    def data_received(self, data):
-        if self.websocket is not None:
-            self.websocket.data_received(data)
-        else:
-            # Pass it to HttpProtocol handler first
-            # That will (hopefully) upgrade it to a websocket.
-            super().data_received(data)
-
-    def eof_received(self) -> Optional[bool]:
-        if self.websocket is not None:
-            return self.websocket.eof_received()
-        else:
-            return False
-
-    def close(self):
-        # Called by HttpProtocol at the end of connection_task
-        # If we've upgraded to websocket, we do our own closure
-        if self.websocket is not None:
-            self.websocket.fail_connection(1001)
-        else:
-            super().close()
-
-    def close_if_idle(self):
-        # Called by Sanic Server when shutting down
-        # If we've upgraded to websocket, shut it down
-        if self.websocket is not None:
-            if self.websocket.connection.state in (CLOSING, CLOSED):
-                return True
-            else:
-                return self.websocket.fail_connection(1001)
-        else:
-            return super().close_if_idle()
-
-    async def websocket_handshake(self, request, subprotocols=Optional[Sequence[str]]):
-        # let the websockets package do the handshake with the client
-        headers = {"Upgrade": "websocket", "Connection": "Upgrade"}
-        try:
-            if subprotocols is not None:
-                # subprotocols can be a set or frozenset, but ServerConnection needs a list
-                subprotocols = list(subprotocols)
-            ws_server = ServerConnection(max_size=self.websocket_max_size, subprotocols=subprotocols,
-                                         state=OPEN, logger=error_logger)
-            resp = ws_server.accept(request)  # type: websockets.http11.Response
-        except Exception as exc:
-            msg = (
-                    "Failed to open a WebSocket connection.\n"
-                    "See server log for more information.\n"
-                )
-            raise SanicException(msg, status_code=500)
-        if 100 <= resp.status_code <= 299:
-            rbytes = b"".join([b"HTTP/1.1 ", b'%d' % resp.status_code, b" ", resp.reason_phrase.encode("utf-8"), b"\r\n"])
-            for k, v in resp.headers.items():
-                rbytes += k.encode("utf-8") + b": " + v.encode("utf-8") + b"\r\n"
-            if resp.body:
-                rbytes += b"\r\n" + resp.body + b"\r\n"
-            rbytes += b"\r\n"
-            await super().send(rbytes)
-        else:
-            raise SanicException(resp.body, resp.status_code)
-
-        self.websocket = WebsocketImplProtocol(ws_server, ping_interval=self.websocket_ping_interval, ping_timeout=self.websocket_ping_timeout)
-        loop = request.transport.loop if hasattr(request, "transport") and hasattr(request.transport, "loop") else None
-        await self.websocket.connection_made(self, loop=loop)
-        return self.websocket
- 
-
-class WebSocketConnection:
-    """
-    This is for ASGI Connections.
-    It provides an interface similar to WebsocketProtocol, but
-    sends/receives over an ASGI connection.
-    """
-    # TODO
-    # - Implement ping/pong
-
-    def __init__(
-        self,
-        send: Callable[[ASIMessage], Awaitable[None]],
-        receive: Callable[[], Awaitable[ASIMessage]],
-        subprotocols: Optional[List[str]] = None,
-    ) -> None:
-        self._send = send
-        self._receive = receive
-        self._subprotocols = subprotocols or []
-
-    async def send(self, data: Union[str, bytes], *args, **kwargs) -> None:
-        message: Dict[str, Union[str, bytes]] = {"type": "websocket.send"}
-
-        if isinstance(data, bytes):
-            message.update({"bytes": data})
-        else:
-            message.update({"text": str(data)})
-
-        await self._send(message)
-
-    async def recv(self, *args, **kwargs) -> Optional[str]:
-        message = await self._receive()
-
-        if message["type"] == "websocket.receive":
-            return message["text"]
-        elif message["type"] == "websocket.disconnect":
-            pass
-
-        return None
-
-    receive = recv
-
-    async def accept(self, subprotocols: Optional[List[str]] = None) -> None:
-        subprotocol = None
-        if subprotocols:
-            for subp in subprotocols:
-                if subp in self.subprotocols:
-                    subprotocol = subp
-                    break
-
-        await self._send(
-            {
-                "type": "websocket.accept",
-                "subprotocol": subprotocol,
-            }
-        )
-
-    async def close(self, code: int = 1000, reason: str = "") -> None:
-        pass
-
-    @property
-    def subprotocols(self):
-        return self._subprotocols
-
-    @subprotocols.setter
-    def subprotocols(self, subprotocols: Optional[List[str]] = None):
-        self._subprotocols = subprotocols or []
